Localise all references to 'mobile'
Remove the mobile property and set it locally in each location
where it is used (preparation for iterating the different
platform+quality combinations which affect the graphics).

Mark each point where a decision is made based on platform+quality
with a plain English version of the combinations which take
effect and when.

Add an explicit hud property which is set from the start screen
and which is independent of the platform.