Be more discriminating when handling touch events
Check the position and type of the touch event to determine
whether acceleration is being applied.

The only issue with this is that if a touch moves from the right
of the screen to the left, if a touchend event occurs in the left
side of the screen, it won't stop the acceleration. But it's a
better solution than counting the number of touches.

Also added a check on touch move events, so that only moves which
occur on the left of the screen are counted as turns. The reason
for this is that in some cases, the touch identifier was being
incorrectly applied: so a touch which started on the left of
the screen could accidentally be continued by a touchmove
occurring on the right of the screen, resulting in a violent right
turn. (It appears as though the touch ID can mistakenly
transfer between touches with different fingers.)

By testing where the touchmove occurs, the violent turn effect
can be removed, as only a finger on the left-hand side of the
screen can cause moves.